chore: add RootLayoutBase component
