+++ src/app/(full-width)/layout.tsx
@@ -1,7 +1,8 @@
-import { RootLayoutBase } from '../layout';
+import RootLayoutBase from '../components/RootLayoutBase';
 
 export default function Layout({
   children,
 }: Readonly<{ children: React.ReactNode }>) {
   return <RootLayoutBase>{children}</RootLayoutBase>;
 }
+
